Extract format-specific index production from the cursor writer
Pure refactor: SSTableCursorWriter's index machinery — promoted index
blocks serialized incrementally, the tail-counted promotion decision,
Index.db entries, bloom filter and summary appends — moves verbatim
into BigCursorIndexWriter behind a new CursorIndexWriter seam
(startPartition / rowWritten / staticRowWritten / endPartition / block
clock).

The seam is event-shaped because unfiltered descriptors are transient:
implementations needing block-boundary clusterings must capture them
at row time, which is also how the BIG code already worked (eager
first-clustering serialization, reusable last-clustering copy). One
deliberate substitution: the tail block passes DeletionTime.LIVE where
the old code read the writer's open marker — equivalent because range
tombstones close before partition end.

No behavior change: this branch's full differential suite passes
byte-identically before and after. Kept purely as enabling
infrastructure — later fixes (2GiB index-offset widening, disabled
bloom filter tolerance) are written against BigCursorIndexWriter's
class structure rather than the original monolithic writer.
